+++ src/runtime/contrib/tflite/tflite_runtime.cc
@@ -154,7 +154,7 @@ PackedFunc TFLiteRuntime::GetFunction(const String& name, const ObjectPtr<Object
   // Return member functions during query.
   if (name == "set_input") {
     return PackedFunc([sptr_to_self, this](TVMArgs args, TVMRetValue* rv) {
-      int in_idx = args[0];
+      int in_idx = args[0].cast<int>();
       ICHECK_GE(in_idx, 0);
       this->SetInput(in_idx, args[1]);
     });
@@ -165,7 +165,7 @@ PackedFunc TFLiteRuntime::GetFunction(const String& name, const ObjectPtr<Object
     return PackedFunc([sptr_to_self, this](TVMArgs args, TVMRetValue* rv) { this->Invoke(); });
   } else if (name == "set_num_threads") {
     return PackedFunc([sptr_to_self, this](TVMArgs args, TVMRetValue* rv) {
-      int num_threads = args[0];
+      int num_threads = args[0].cast<int>();
       CHECK_GE(num_threads, 1);
       this->SetNumThreads(num_threads);
     });
